Move delay link into another file
